Add logging module and update CLI loglevel setting
Logging is provided by the python `logging` module and is controlled
via the `--verbose` and `--debug` command line arguments that decrease
the logging level from its default (`WARNING`) to `INFO` or `DEBUG`, if
applied. It is irrelevant where in the command line argument string the
options are set and also, if they are specified multiple times.